add unit

+++ Engine/Collider/BoxCollider2D.cs
@@ -13,15 +13,15 @@ public class BoxCollider2D : Collider2D
     {
         if (other is BoxCollider2D box)
         {
-            float leftA = Position.X;
-            float rightA = Position.X + Size.X;
-            float topA = Position.Y;
-            float bottomA = Position.Y + Size.Y;
+            double leftA = Position.X;
+            double rightA = Position.X + Size.X;
+            double topA = Position.Y;
+            double bottomA = Position.Y + Size.Y;
 
-            float leftB = box.Position.X;
-            float rightB = box.Position.X + box.Size.X;
-            float topB = box.Position.Y;
-            float bottomB = box.Position.Y + box.Size.Y;
+            double leftB = box.Position.X;
+            double rightB = box.Position.X + box.Size.X;
+            double topB = box.Position.Y;
+            double bottomB = box.Position.Y + box.Size.Y;
 
             // Проверяем, не находятся ли коллайдеры вне друг друга
             bool noCollision = rightA <= leftB || // A полностью слева от B
